requirements.txt
fastapi
uvicorn
pydantic
nltk
python-dotenv
scikit-learn
joblib
groq
requests